Now, the CG works but the model is not good.

+++ src/iteratorscustom.rs
@@ -1,14 +1,14 @@
 //! Create a float iterator given the `start`, `stop` and the number of elements `steps`
 pub struct FloatIterator {
-    current: u64,
-    current_back: u64,
-    steps: u64,
-    start: f64,
-    end: f64,
+    current: u32,
+    current_back: u32,
+    steps: u32,
+    start: f32,
+    end: f32,
 }
 
 impl FloatIterator {
-    pub fn new(start: f64, end: f64, steps: u64) -> Self {
+    pub fn new(start: f32, end: f32, steps: u32) -> Self {
         FloatIterator {
             current: 0,
             current_back: steps,
@@ -18,24 +18,24 @@ impl FloatIterator {
         }
     }
 
-    pub fn length(&self) -> u64 {
+    pub fn length(&self) -> u32 {
         self.current - self.current_back
     }
 
-    pub fn at(&self, pos: u64) -> f64 {
-        let f_pos = pos as f64 / self.steps as f64;
+    pub fn at(&self, pos: u32) -> f32 {
+        let f_pos = pos as f32 / self.steps as f32;
         (1. - f_pos) * self.start + f_pos * self.end
     }
 
     fn usize_len(&self) -> usize {
         let l = self.length();
-        debug_assert!(l <= ::std::usize::MAX as u64);
+        debug_assert!(l <= ::std::usize::MAX as u32);
         l as usize
     }
 }
 
 impl Iterator for FloatIterator {
-    type Item = f64;
+    type Item = f32;
 
     fn next(&mut self) -> Option<Self::Item> {
         if self.current >= self.current_back {
